Added 0.22.54 objects
OpenSteel/OS 0.22 Denver Build 54 is not the final Circuit 4 build, but it implements %d and %x into printf(char* str, ...) and adds an itoa() function to libstr. It includes from five integers to ensure it is working as intended. Also, despite my hesitance to do so, I have added a GCC header (not inside the actual code, but as an include), because there's simply no pragmatic way to implement a decent va_arg and va_list and whatnot on a complete self-reliant model. I have tested it multiple times, and it should compile and build a binary without issue.

This build may not be completely free from bugs, but it'll be fine, as long as you're not an avid diehard fan of percentage symbols and use OpenSteel/OS to create screens full of them. You probably found a manual workaround to the Shift+5 bug already, and that displays just fine still, so...