Updated speaker.h
The struct will be discarded quickly (for now) and placed into a special index archive of sorts (on-spot terminology :P), as it doesn't work and for now, I'll opt for the use of BeepVariable. In order to use a musical note, you'll have to use the equation f = 2^((note - 69) / 12) * 440. This is regardless of if there was a struct or not, it's simple if you have a TI-84 or Casio FX-300ES or some other calc, just plug the note (like 60) into where I put note. Git commit messages don't let me input properly formatted equations.