Added 0.22.70 objects
Currently in the midst of a 16-computer hardware test of the previous main build, which I opted not to upload due to... reasons.

This build features a split up interrupts.cpp, which while bloating the binary and ISO fileisze by a few kilobytes, should streamline this aspect a whole lot more. A style guide also has been adopted.