Updated a few files
 - Updated the assembly index.
 - Added the current code written for NRSh (Nathan Renaud Shell, ofc we gotta honour the narrator every now and then > _ l)
 - Added code which now includes panic() and 100% less PrintfKeyboardEventHandler (why am i treating this like its a good thing, it'll probably come back to bite me in the ass one day, likely sooner rather than later with my development of a shell).

Additional note: I realised a lot of my keyboard driver is implemented in a horrendous, horrible, rancid, terrible, disgusting, [insert 52 more adjectives which describe something in a negative way] design. By implementing it the way I did, it means OpenSteel/OS thinks everyone has a QWERTY keyboard as they are found in the United States and Canada. I'll work towards a solution for this in the future, but it isn't a top concern.