OpenSteel/OS 0.22 Build 153
So, there's a big development in OpenSteel/OS 0.22 Build 153. I know I said the last thing would be a serial port driver, and that's still on my radar, but I also opted to write a command line since I couldn't wait much longer. There's currently four commands: ver, help, echo and cls. I also reverted the background colour back to black, mainly to reduce the amount of blue light produced by the OS and to also be nice on your eyes.

I should note that there are some issues with the shell, mainly that you can still use the backspace key normally and delete the " > " bit, and that there's no text cursor. I'll spend the next few days writing new commands, testing them all, and resolving the current oversights. Making it impossible to delete the " > " should be as easy as implementing a safeguard.

Anyways, as per usual, open an issue if you catch a bug; provide a screenshot to help provide better understanding.