Updated to kernel 0.22.155
I said something about implementing a guard to prevent the " > " from being removed by backspace earlier. I did two things:
 (1) Changed " > " to "> ", as I am opting to remove the 1 char padding on both the left and right.
 (2) Moved the printf("\b"); from keyboard.cpp's PrintAndNotify mode to nrshManager. It has been retained in PrintOnly.

Doing this, in my testing, fixed this oversight. And one of my main concerns was that it'd cause an underflow in inputLength, but (1) it's unsigned and so negative numbers are available, and (2) nrshManager already knew what to do in that case if " > " was typed out. It's almost 10 PM for me, give me credit where it's due that I was able to fix something (at least a patch in my eyes) quickly and competently while sitting in bed with my laptop over a blanket (T480 with the chonk battery, it gets adequate airflow as long as you're not compiling something like the Linux kernel entirely in one lying).

I'll update the keyboard driver, version header and build objects shortly. I'll focus on getting a textmode cursor a la MS-DOS into OpenSteel/OS tomorrow, maybe there'll be a round of commits but I'm putting a 2 hour limit on OS dev over the Thanksgiving break so that I have time to work on studying and video production.