OpenSteel/OS 0.22 "Denver" Build 159 (14:04:28 2025-11-26)
This build features some minor adjustments to the Nathan Renaud Shell (NRSh), the kernel appearance, and the keyboard driver.

As I stated last night with OpenSteel/OS 0.22 Build 155, I implemented a safeguard for NRSh to stop "> " from being typed out, but in the process, I made it impossible to use the backspace key in the PrintAndNotify keyboard mode. Build 159 rectifies that issue, by using the bool isShellAvailable to determine if NRSh is active. This means the logic for the backspace key is slightly more awkward, but it works. I also removed the 1-char padding on both the left and right sides of the screen, as while it looked nice and whatnot, it also has pragmatic issues that'd involve video memory adjustments, and pragmatism should come before personal aesthetic choices.

I also made some colouring adjustments to the "> " bit, so now it's an aqua colour instead of white, in addition to making the first steps towards not needing to type the enter key to enable NRSh. It's flawed, but now it is part of the way there. I'll open an issue in the meantime. Anyways, I'm going to go get some lunch, and work on some other things after that. The textmode cursor is coming, but give me some time.