Removed nrshManager() in kernel.cpp
Replaced it with a more functional solution. It's a bit rancid in relation to multitasking, but that wasn't really present in any previous build. I'm NOT getting rid of multitasking, it's just that to utilise NRSh right now, the kernel's while(1) loop has to be used. It's rancid, but it'll work for NOW. Probably not in a MONTH.