OpenSteel/OS 0.22 "Denver" Build 161 (2025-11-26 23:50:52)
This build rectifies the previous issue with the keyboard driver <-> NRSh <-> kernel conflict (mostly). There is an issue for now, in that any multitasking is essentially crippled, but it already was when NRSh was implemented. It also fixes the bug I opened earlier and self-assigned where the first input prompt would be completely ignored.

I still don't have the text mode cursor, or any additional commands, but I guess that's a day of decent development. With US Thanksgiving Day arriving in minutes, I'm going to be visiting family (in addition to a family member's friend coming with us), and so I hope to show my project to relatives for that little tiny bit of "showoff" I rarely release.

Anyways, I gotta get sleep and edit audio tomorrow, good night.