OpenSteel/OS 0.22 "Denver" Build 164 (2025-11-27 18:09:34)
Again, like 0.22.161, this build is a minor quality-of-life adjustment to its predecessor. This time, the boolean isShellAvailable has been deprecated in favour of an enumeration, InterfaceModes, which includes multiple types. The two main ones of concern right now are BootUI and TextShell.

I'll focus next on adding some additional commands for credits, licensing and shutting down, as I previously stated I'd do in my log. But, as usual, let me know of any major issues you experience so that I can try to fix them.