OpenSteel/OS 0.22 "Denver" Build 170 (2025-11-27 20:57:53)
The only new thing in this build is that I added the code for the credits command.

So why did we go right from 0.22.166 to 0.22.170? Well, blame the rancid obsessions of gen alpha for this build not being 0.22.167 - If it was compiled last year, what this is, 0.22.170, would be 0.22.167, but I don't want the braindead attentionspanless gen alpha fools to start sticking their tongues out at specific builds of my project. Yes, I know that it doesn't help by bringing attention to it. Look, just keep quiet about this, okay?