OpenSteel/OS 0.22 "Denver" Build 182 (2025-12-01 17:12:54)
New month, new build only 17 hours in.

This build introduces (publicly) CPUID functions/logic that allows for OpenSteel/OS to make a good guess as to the CPU inside your computer. There are cases in which this may not work and the strings from CPUID are bogus (like emulating the i486 thru Pentium III in QEMU, something I tested myself), but it should be fine on most modern systems, both Intel and AMD. This might be used as a backend to determine if SSE and SSE2 and whatnot are available and can be utilised in the future, but OpenSteel/OS doesn't support those technologies yet.

In order to see this in action, you can enter the command cpudata from the NRSh command prompt. On the topic of that, I feel that with so many commands being added to NRSh, the current help system is going to falter quickly with giving each command its own line, as in the rare event that I add 22+ commands total (chances are very low, but never zero, given my autism that this project flourishes because of, like many older projects), the help command will buckle. And so, command descriptions and examples of use will likely be ported over to a "-?" flag in each command, but the libraries and logic to allow this still needs to be written, hence the currently empty strtok I'm going to implement that is sitting in strlib.cpp.

Anyways, I've gotta do other things, like work on algebraic things, so as per usual, report any bugs, and enjoy this build.