Updated libstr.cpp (again)
Because strtok worked like a pre-2022 steelsofliquid computer: horribly, that i'm impressed it works. I would've made another comparison, but it was fucked up, and edgy humour is for my property (i.e. room and offline storage media) only, not publicly seen git commit messages.

And so strtok had to be updated.