Patched echo in NRSh (coup de grace, for now)
Issue #15 is now resolved, although this second patch is mostly unrelated.
This alters some logic in NRSh's core, allowing for 32 tokens (or up to 31 arguments/flags) to be parsed (executed). The previous number was 8.

I could've done this all (and if my brain remembers correctly at this hour, did part of) shortly after 0.22.191 was published, if it weren't for dinner and a shower.