Updated NRSh code
(kernel.cpp in here as I need to remove the build 200 celebration code)

I added a small beep comparable to the IBM PC AT in the PCjs emulator (a website I've enjoyed using for many years). Specifically, a beep I presume is close to, if not, 741 Hz, which will play when there's nothing to backspace in NRSh. The speaker driver construction in NRSh is isolated in a bid to not interfere with the already constructed driver in the kernel.

I plan to add more system sounds in the near future.