OpenSteel/OS 0.22 "Denver" Build 202 (2025-12-03 21:46:47)
This build introduces a minor change in the forme of a short 1-tick beep that is played when backspacing when there isn't anything to backspace.

Just a little thing I thought would be easy to do, and it was easy. But outside of NRSh and the kernel, the speaker driver won't be used, at least not until I can develop either a clean system call or clean API for it.